Introduced USE_BLOSC profile in maven, this allwos skipping blosc tests if blosc library is not available and not configured
- renamed maven.properties to blosc.properties, and the template too
- moved blosc tests into seperate class
- updated Readme.md
- updated .gitignore as necessary
