+++ template.blosc.properties
@@ -1,4 +1,4 @@
-# Make a copy of this file in the same directory and rename it from "template.maven.properties" to "maven.properties".
+# Make a copy of this file in the same directory and rename it from "template.blosc.properties" to "blosc.properties".
 # Replace "<...>" marked values of properties with real values.
 
 # On Windows the library is not part of the system. It has to be compiled.
